// types/experiment.d.ts
export interface ExperimentAPIResponse {
  id: string;
  dateCreated: string;
  dateUpdated: string;
  name: string;
  project: string;
  hypothesis: string;
  description: string;
  tags: string[];
  owner: string;
  archived: boolean;
  status: string;
  autoRefresh: boolean;
  hashAttribute: string;
  fallbackAttribute?: string;
  hashVersion: 1 | 2;
  disableStickyBucketing?: any;
  bucketVersion?: number;
  minBucketVersion?: number;
  variations: Array<{
    variationId: string;
    key: string;
    name: string;
    description: string;
    screenshots: string[];
  }>;
  phases: Array<{
    name: string;
    dateStarted: string;
    dateEnded: string;
    reasonForStopping: string;
    seed: string;
    coverage: number;
    trafficSplit: Array<{
      variationId: string;
      weight: number;
    }>;
    namespace?: {
      namespaceId: string;
      range: [number, number];
    };
    targetingCondition: string;
    savedGroupTargeting?: Array<{
      matchType: "all" | "any" | "none";
      savedGroups: string[];
    }>;
  }>;
  settings: {
    datasourceId: string;
    assignmentQueryId: string;
    experimentId: string;
    segmentId: string;
    queryFilter: string;
    inProgressConversions: "include" | "exclude";
    attributionModel: "firstExposure" | "experimentDuration";
    statsEngine: "bayesian" | "frequentist";
    regressionAdjustmentEnabled?: boolean;
    goals: Array<{
      metricId: string;
      overrides: {
        delayHours?: number;
        windowHours?: number;
        window?: "conversion" | "lookback" | "";
        winRiskThreshold?: number;
        loseRiskThreshold?: number;
      };
    }>;
    secondaryMetrics: Array<{
      metricId: string;
      overrides: {
        delayHours?: number;
        windowHours?: number;
        window?: "conversion" | "lookback" | "";
        winRiskThreshold?: number;
        loseRiskThreshold?: number;
      };
    }>;
    guardrails: Array<{
      metricId: string;
      overrides: {
        delayHours?: number;
        windowHours?: number;
        window?: "conversion" | "lookback" | "";
        winRiskThreshold?: number;
        loseRiskThreshold?: number;
      };
    }>;
    activationMetric?: {
      metricId: string;
      overrides: {
        delayHours?: number;
        windowHours?: number;
        window?: "conversion" | "lookback" | "";
        winRiskThreshold?: number;
        loseRiskThreshold?: number;
      };
    };
  };
  resultSummary?: {
    status: string;
    winner: string;
    conclusions: string;
    releasedVariationId: string;
    excludeFromPayload: boolean;
  };
}